+++ crates/azure/src/role_assignments.rs
@@ -6,6 +6,9 @@ use std::path::PathBuf;
 use std::time::Duration;
 use tracing::debug;
 
+/// Fetches all AzureRM role assignments.
+/// 
+/// Not to be confused with Entra role assignments.
 pub async fn fetch_all_role_assignments() -> Result<Vec<RoleAssignment>> {
     debug!("Fetching role assignments");
     let mut query = ResourceGraphHelper::new(