+++ crates/azure_types/src/access_token.rs
@@ -1,0 +1,35 @@
+use std::fmt::Debug;
+
+use chrono::{DateTime, Local};
+use serde::{Deserialize, Serialize};
+
+use crate::{prelude::SubscriptionId, tenants::TenantId};
+
+#[derive(Deserialize)]
+pub struct AccessToken<T> {
+    #[serde(rename="accessToken")]
+    pub access_token: T,
+    #[serde(deserialize_with = "crate::serde_helpers::deserialize_local_date_time_from_epoch")]
+    pub expires_on: DateTime<Local>,
+    pub subscription: SubscriptionId,
+    pub tenant: TenantId,
+    #[serde(rename="tokenType")]
+    pub token_type: TokenType,
+}
+
+impl<T> Debug for AccessToken<T> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("AccessToken")
+            .field("access_token", &"***redacted***")
+            .field("expires_on", &self.expires_on)
+            .field("subscription", &self.subscription)
+            .field("tenant", &self.tenant)
+            .field("token_type", &self.token_type)
+            .finish()
+    }
+}
+
+#[derive(Deserialize, Serialize, Debug, Clone, Copy, PartialEq, Eq)]
+pub enum TokenType {
+    Bearer,
+}