+++ crates/azure_types/src/unified_role_definition_id.rs
@@ -1,0 +1,4 @@
+#[derive(Debug, Clone, Copy, Eq, PartialEq, Hash)]
+pub struct UnifiedRoleDefinitionId(uuid::Uuid);
+
+crate::impl_uuid_newtype!(UnifiedRoleDefinitionId);