+++ crates/credentials/src/azure_devops_pat.rs
@@ -1,7 +1,7 @@
+use crate::AuthBearerExt;
 #[cfg(windows)]
 use crate::read_azure_devops_pat_from_credential_manager;
 use reqwest::header::HeaderValue;
-use std::io::Write;
 use std::ops::Deref;
 
 #[derive(Debug)]
@@ -41,25 +41,9 @@ impl<'de> serde::Deserialize<'de> for AzureDevOpsPersonalAccessToken {
         Ok(Self::new(value))
     }
 }
-impl AzureDevOpsPersonalAccessToken {
-    pub fn into_authorization_header_value(&self) -> HeaderValue {
-        let mut buf = b"Basic ".to_vec();
-        {
-            let username = "";
-            let password = self;
-            let mut encoder =
-                base64::write::EncoderWriter::new(&mut buf, &base64::prelude::BASE64_STANDARD);
-            let _ = write!(encoder, "{username}:");
-            let _ = write!(encoder, "{password}");
-        }
-        let mut header = HeaderValue::from_bytes(&buf).expect("base64 is always valid HeaderValue");
-        header.set_sensitive(true);
-        header
-    }
-}
 impl From<AzureDevOpsPersonalAccessToken> for HeaderValue {
     fn from(value: AzureDevOpsPersonalAccessToken) -> Self {
-        value.into_authorization_header_value()
+        value.as_authorization_header_value()
     }
 }
 
@@ -77,10 +61,13 @@ mod test {
     use crate::get_azure_devops_pat;
 
     #[tokio::test]
-    #[ignore]
     pub async fn non_empty() -> eyre::Result<()> {
-        let credential = get_azure_devops_pat().await?;
-        assert!(!credential.is_empty());
+        let credential = get_azure_devops_pat().await.ok();
+        if let Some(credential) = credential {
+            assert!(!credential.is_empty());
+        } else {
+            println!("No credential found, skipping test");
+        }
         Ok(())
     }
 }