+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 - Impl `Ord` for `KeyVaultId`
 - Added browse storage accounts action
 - Add `GovernanceRoleAssignmentMemberType::Direct` variant
+- Add `PrincipalCollection` type and change `fetch_all_principals` to return it
+- Add `UnifiedRoleDefinition` and `UnifiedRoleAssignment` types for Entra RBAC
+- Move `AccessToken` type from `cloud_terrastodon_credentials` to `cloud_terrastodon_azure_types`
+- Add jwt decoding to `cloud_terrastodon_credentials` (TODO: fix fn return instead of just printing)
+
 
 # v0.26.0